Add support for specifying column type.
Instead of having to set the whole column, we can just override the type
of the field itself.

Additionally, we allow fields to have a custom type so that it's easier
to just have blanket support for conversion of types without having to
specify on specific fields.

This works much nicer with the rest of sqlmodel.