:bug: Fix enum type checks ordering in `get_sqlalchemy_type`
When enum subclassing/mix-in[1] is used, especially to ensure compatibility
and produce correct OpenAPI specifications with FastAPI[2], SQLModel on its
side doesn't map to the correct SQLAlchemy type.

e.g.:
* `class Foo(Enum)` will work but won't allow a proper OpenAPI spec. to be
generated.
* `class Foo(str, Enum)` will fix the previous situation, but prevents
SQLModel to read and map the object from the DB (due to validation errors).

Changing ordering of checks to ensure that Enum subclasses are tested first.

[1] https://docs.python.org/3/library/enum.html#restricted-enum-subclassing
[2] https://fastapi.tiangolo.com/sq/tutorial/path-params/#create-an-enum-class