Add config option to load relationship fields.
Due to the recursive loading problem loading relations is not yet possible.
This change introduces the config option 'include_relations' to also load specifically chosen relations.

For the current use case nothing changes unless the user specifically sets fields to be included (and carefully considers the risks of circular includes).
The option is very valuable if the table design contains many 1:n relations.